OOGA-57: Added new API abstractions and methods

+++ src/main/java/oogasalad/editor/controller/api/HitboxDataAPIInterface.java
@@ -3,15 +3,15 @@ package oogasalad.editor.controller.api;
 import java.util.UUID;
 
 public interface HitboxDataAPIInterface {
-  public double getX(UUID id);
-  public double getY(UUID id);
-  public double getWidth(UUID id);
-  public double getHeight(UUID id);
+  public int getX(UUID id);
+  public int getY(UUID id);
+  public int getWidth(UUID id);
+  public int getHeight(UUID id);
   public String getShape(UUID id);
 
-  public void setX(UUID id, double x);
-  public void setY(UUID id, double y);
-  public void setWidth(UUID id, double width);
-  public void setHeight(UUID id, double height);
+  public void setX(UUID id, int x);
+  public void setY(UUID id, int y);
+  public void setWidth(UUID id, int width);
+  public void setHeight(UUID id, int height);
   public void setShape(UUID id, String shape);
 }
